requirements.txt
ipython
azureml-sdk
